requirements.txt
alembic==1.16.2
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.4
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
eventkit==1.0.3
Flask==3.1.1
fonttools==4.58.5
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
hmmlearn==0.3.3
ib-insync==0.9.86
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lightgbm==4.6.0
loguru==0.7.3
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
multitasking==0.0.11
narwhals==1.46.0
nest-asyncio==1.6.0
numpy==1.26.4
optuna==4.4.0
optuna-dashboard==0.19.0
packaging==25.0
pandas==2.3.1
pandas-datareader==0.10.0
pandas_ta==0.3.14b0
peewee==3.18.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.46.1
streamlit-cookies-manager
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
xlsxwriter==3.2.5
yfinance==0.2.64
pytest==8.2.2
